+++ data-transfer-object/src/main/java/org/grnet/cat/dtos/registry/motivation/MotivationResponse.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
-import org.grnet.cat.dtos.registry.MotivationActorPartialResponse;
+import org.grnet.cat.dtos.registry.actor.MotivationActorPartialResponse;
 
 import java.util.Arrays;
 import java.util.List;